Multiple interpretations issue of the output argument
The initiative here is to use the value of the output argument only for
the naming of output directories. It is currently also used for the
naming of arguments files, but this usage is not in line with the
description of this argument given by the wiki of the simulator.