Regarding the CombinedGui and ConfigurationGui interplay
This interplay is altered by having the ConfigurationGui set up a thread
to run the EvolutionGui each time an evolution is to be run. This
eliminates the need for the thread in the CombinedGui that was designed
to wait for a new evolution in order to run the EvolutionGui. As a
result, the configuration results now get pushed into the CombinedGui
where they previously were being pulled into. This allows for more
structured and cleaner code.